requirements.txt
# Flask web framework
Flask==2.3.3
Werkzeug==2.3.7

# Document parsing
python-docx==0.8.11
pdfplumber==0.10.3
python-magic==0.4.27

# Data processing
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
XlsxWriter==3.1.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Other utilities
Pillow==10.0.0
